test: start refactoring to use Hilt for tests
Hilt requires some setup for the test files, being refactoring them all to work with it, for now the
Activity tests work, and the other files are currently just commented out to avoid compilation
issues.
