+++ app/build.gradle
@@ -122,7 +122,7 @@ dependencies {
     androidTestImplementation "androidx.test.espresso:espresso-intents:3.3.0"
     androidTestImplementation "androidx.test.ext:junit:1.1.2"
     androidTestImplementation "androidx.test:rules:1.3.0"
-    // androidTestImplementation "androidx.work:work-testing:2.0.1"
+    androidTestImplementation "com.google.dagger:hilt-android-testing:$hilt_version"
     androidTestImplementation("com.schibsted.spain:barista:3.7.0") {
         exclude group: "com.android.support"
         exclude group: "org.jetbrains.kotlin"
@@ -131,8 +131,7 @@ dependencies {
 
     kapt "androidx.hilt:hilt-compiler:$hiltx_version"
     kapt "com.google.dagger:hilt-android-compiler:$hilt_version"
-    kaptAndroidTest "com.google.dagger:dagger-android-processor:$dagger_version"
-    kaptAndroidTest "com.google.dagger:dagger-compiler:$dagger_version"
+    kaptAndroidTest "com.google.dagger:hilt-android-compiler:$hilt_version"
 }
 
 kapt {
