+++ app/src/androidTest/java/com/chesire/nekome/TestRunner.kt
@@ -3,9 +3,10 @@ package com.chesire.nekome
 import android.app.Application
 import android.content.Context
 import androidx.test.runner.AndroidJUnitRunner
+import dagger.hilt.android.testing.HiltTestApplication
 
 /**
- * Overridden runner that forces the application object to be [com.chesire.nekome.TestApplication].
+ * Overridden runner that forces the application object to be [HiltTestApplication].
  */
 @Suppress("unused")
 class TestRunner : AndroidJUnitRunner() {
@@ -13,5 +14,5 @@ class TestRunner : AndroidJUnitRunner() {
         cl: ClassLoader?,
         className: String?,
         context: Context?
-    ): Application = super.newApplication(cl, "com.chesire.nekome.TestApplication", context)
+    ): Application = super.newApplication(cl, HiltTestApplication::class.java.name, context)
 }
