+++ app/src/androidTest/java/com/chesire/nekome/flow/series/FilterTests.kt
@@ -1,5 +1,5 @@
 package com.chesire.nekome.flow.series
-
+/*
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isChecked
 import androidx.test.espresso.matcher.ViewMatchers.isNotChecked
@@ -127,3 +127,4 @@ class FilterTests {
         )
     }
 }
+*/
