+++ app/src/androidTest/java/com/chesire/nekome/flow/series/SeriesListTests.kt
@@ -1,5 +1,5 @@
 package com.chesire.nekome.flow.series
-
+/*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
 import com.chesire.nekome.Activity
@@ -44,3 +44,4 @@ class SeriesListTests {
         assertDisplayed(R.id.listEmpty)
     }
 }
+*/
