+++ app/src/androidTest/java/com/chesire/nekome/flow/series/SeriesViewHolderTests.kt
@@ -1,5 +1,5 @@
 package com.chesire.nekome.flow.series
-
+/*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
 import com.chesire.nekome.Activity
@@ -233,3 +233,4 @@ class SeriesViewHolderTests {
         )
     }
 }
+*/
