+++ app/src/androidTest/java/com/chesire/nekome/flow/series/SortTests.kt
@@ -1,5 +1,5 @@
 package com.chesire.nekome.flow.series
-
+/*
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.rule.ActivityTestRule
 import com.chesire.nekome.Activity
@@ -52,3 +52,4 @@ class SortTests {
         assertNotExist(R.string.sort_dialog_title)
     }
 }
+*/
