--- app/src/androidTest/java/com/chesire/nekome/helpers/MethodHelpers.kt
@@ -1,19 +1,0 @@
-package com.chesire.nekome.helpers
-
-import androidx.test.platform.app.InstrumentationRegistry
-import com.chesire.nekome.TestApplication
-import com.chesire.nekome.injection.components.TestComponent
-
-/**
- * Gets the test context as [TestApplication].
- */
-fun getApplication() = InstrumentationRegistry
-    .getInstrumentation()
-    .targetContext
-    .applicationContext as TestApplication
-
-/**
- * Gets the required component to inject test files.
- */
-val injector: TestComponent
-    get() = getApplication().component
