+++ app/src/androidTest/java/com/chesire/nekome/injection/FakeUrlModule.kt
@@ -1,13 +1,18 @@
-package com.chesire.nekome.injection.modules
+package com.chesire.nekome.injection
 
+/*
 import com.chesire.nekome.core.url.UrlHandler
 import com.chesire.nekome.harness.FakeUrlHandler
 import dagger.Binds
 import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ApplicationComponent
 
 @Suppress("unused")
 @Module
+@InstallIn(ApplicationComponent::class)
 abstract class FakeUrlModule {
     @Binds
     abstract fun providesFakeUrlHandler(fakeUrlHandler: FakeUrlHandler): UrlHandler
 }
+*/
